Backend: periodically trigger refresh of all power-adapter states
We track individual changes to the state of the power-port through
backend calls but it may very well be possible that some outside
source like a script also modifies the state.

Therefore sync up our internal state to the real world periodically.
We'll do this simply inside the general power read for the main
supply which will trigger this every 30 seconds.