Power: Servo: use spi_hold as power-indicator
power_state does not return the actual power-state on both my pinky and gru.
Comparing dut-control vars between on and off states showed that instead
spi_hold is one of the few variables that seems a usable indicator for
the current power state - let's see if this holds up for future devices.