Backend: use the actual halt param to shutdown
-h in lower-case nowadays is an alias to poweroff.
The upper-case -H is now the command for shutting down without
powering down the machine.

As the boardfarm-components are connected to controllable
power-supply we really only want them to halt, as powering down
sometimes also causes strange behaviours.

For example the remote puma board is configured to always power-on
and therefore will reboot again after the power-down operation.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>