Power: add an adapter for Qemu emulation
To test kernel changes on a number of different platforms it may be necessary
to also include a "clean" platform in this test scheme, like on RiscV where
not that much hardware is available.

To streamline such a testing process with my available process
(build one kernel, run on multiple machines), add the possibility
to add Qemu-emumlated machines to my boardfarm.

Using daemonize and the process name it's nicely possible to even
stop them from the boardfarm interface.

Thanks to Stefan Koch for documenting this nicely:
https://blog.stefan-koch.name/2020/05/24/managing-qemu-vms

Signed-off-by: Heiko Stuebner <heiko@sntech.de>