Update README with current version expectations
The underlying library versions were never exactly specified
before, but the recent move to upgrade the debian systems below
revealed a number of bit-rotten versions.

So now after having determined the easiest way to keep everything
working in a Debian/bookworm (+bullseye for python2) setting,
document these things in the README.

Also document the way forward, especially of ignoring the 6.x.x
releases and moving directly to 7.x.x when it comes to it.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>